refactor: component 합성을 이용한 코드 재사용
