+++ frontend/src/components/profile/Profile.js
@@ -1,21 +1,10 @@
-import React, { useState, useEffect, useRef } from "react";
+import React from "react";
+import Croucher from "../base/Croucher";
 import UserInfo from "../../data/user/profile.json";
 import "./Profile.css";
 
-const Profile = ({ onSearch }) => {
+export default function Profile({ onSearch }) {
   const { name, id, loc, org } = UserInfo;
-  const [pageY, setPageY] = useState(0);
-  const documentRef = useRef(document);
-  const handleScroll = () => {
-    const { pageYOffset } = window;
-    setPageY(pageYOffset);
-  };
-  useEffect(() => {
-    documentRef.current.addEventListener("scroll", handleScroll);
-    return () =>
-      documentRef.current.removeEventListener("scroll", handleScroll);
-  }, [pageY]);
-
   const handleSearch = idx => {
     /* loc 클릭시 전체글, org 클릭시 해당 부대 글 */
     return () => onSearch({ keyword: { 0: "", 1: org }[idx] });
@@ -31,16 +20,14 @@ const Profile = ({ onSearch }) => {
       <div className="name">
         {name} @{id}
       </div>
-      <div className={pageY <= 180 ? "belong" : "belongOnly"}>
+      <Croucher norm={180} stretch="belong" crouched="belongOnly">
         <button className="loc" onClick={handleSearch(0)}>
           {loc}
         </button>
         <button className="org" onClick={handleSearch(1)}>
           {org}
         </button>
-      </div>
+      </Croucher>
     </div>
   );
-};
-
-export default Profile;
+}
